# spyglass-spring-webmvc/src/test/resources/apidocs-test/openapi-fixture.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fixture Explorer API",
    "version": "1.0.0",
    "x-service-name": "fixtureService"
  },
  "paths": {
...
    "/form": {
      "post": {
        "tags": ["Bodies"],
        "summary": "Submit a form (urlencoded)",
        "description": "An `application/x-www-form-urlencoded` body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "count": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ok" }
        }
      }
    },
...
    "/examples": {
      "post": {
        "tags": ["Bodies"],
        "summary": "Echo a payload (named examples)",
        "description": "Showcases spec-provided named `examples` on the request body, a parameter and the response.",
        "parameters": [
          {
            "name": "filter", "in": "query", "required": false, "schema": { "type": "string" },
            "examples": {
              "by-name": { "summary": "Match by name", "value": "acme" },
              "by-id": { "summary": "Match by id", "value": "42" }
            }
          },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" }, "example": 20 }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExamplePayload" },
              "examples": {
                "minimal": { "summary": "Just the required field", "value": { "name": "widget" } },
                "full": {
                  "summary": "Every field populated",
                  "description": "A fully populated payload.",
                  "value": { "name": "deluxe widget", "count": 7, "tags": ["alpha", "beta"] }
                },
                "external": { "summary": "Loaded from a URL", "externalValue": "https://example.invalid/sample-payload.json" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The echoed payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExamplePayload" },
                "examples": {
                  "created": { "summary": "A freshly created payload", "value": { "name": "widget", "count": 1, "tags": ["new"] } },
                  "archived": { "summary": "An archived payload", "value": { "name": "old widget", "count": 0, "tags": ["archived"] } }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamplePayload": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "The payload name.", "example": "widget" },
          "count": { "type": "integer", "description": "An optional count.", "example": 5 },
          "tags": { "type": "array", "items": {}, "example": ["a", "b"] }
        }
      },
...
    }
  }
}